test: consolidate location tests into raceStageClassifications.test.js mega test
Merge RaceStageLocationPoints and RaceStageLocationMountains tests into
raceStageClassifications.test.js to create unified stage results test:

- Add locationPoints and locationMountains test data to CLASSIFICATIONS_TEST_CASES
- Add 4 new CSV output tests (Location Points headers/content, Location Mountains headers/content)
- Fix classification JSON fixture paths to match actual filenames
- Rename typo in mountain fixture file (,json -> .json)
- Remove separate location test files

All stage-related tests (results, classifications, sprints, KOMs) now in one
mega test file sharing the same HTML fixture.
