+++ test/scraping/cycling/procyclingstats/fixtures/raceStageResults-classification-mountain-2025-tour-down-under-1.json
